charaka_app/requirements.txt
streamlit>=1.32.0
torch>=2.0.0
sentencepiece>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.1.0
pypdf>=4.0.0
pdf2image>=1.17.0
